pyproject.toml
[project]
name = "telecom-assistant-hackathon"
version = "0.1.0"
description = ""
authors = [
    {name = "Rahul"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "langgraph (>=1.0.4,<2.0.0)",
    "langchain (>=1.1.0,<2.0.0)",
    "crewai (>=1.6.1,<2.0.0)",
    "pyautogen (>=0.10.0,<0.11.0)",
    "llama-index (>=0.14.8,<0.15.0)",
    "pandas (<2.3.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "crewai-tools (>=1.6.1,<2.0.0)",
    "langchain-openai (>=1.1.0,<2.0.0)",
    "faiss-cpu (>=1.13.0,<2.0.0)",
    "llama-index-vector-stores-faiss (>=0.5.1,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
